requirements.txt
Jinja2==3.1.4
PyYAML==6.0.2
pre-commit==4.0.1
python-dotenv==1.0.1